app_flask_eval_pdf_email/requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
docx2pdf==0.1.8
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
reportlab==4.4.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
